Complete backend integration and match features

+++ back/src/models/user.js
@@ -35,9 +35,9 @@ const User = sequelize.define("User",{
             defaultValue:"en",
         },
         points:{
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(10, 2),
             allowNull:false,
-            defaultValue:0,
+            defaultValue:0.00,
         },
     },
     {
